- altered css layout to give app a component feel - updated default states - updated bar weight based on metric toggle - redeveloped input - added metric toggle - separated weightInput css by major components and began utilizing scss
