+++ src/App.css
@@ -3,22 +3,17 @@
   }
 
   .weightDisplay {
-    margin-left: auto;
-    margin-right: auto;
-    padding-bottom: 1em;
-    width: 90%;
-    /* border: 5px dotted lightgray; */
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 300px;
+    margin: 5px;
+    display: grid;
+    grid-auto-flow: column;
+    grid-template-columns: 1fr;
+    overflow: hidden;
 }
 
   .content {
+    width: 555px;
     display: flex;
-    min-height: 100vh;
     flex-direction: column;
-    justify-content: center;
     border: solid;
     border-radius: 1rem;
   }
