+++ src/components/WeightInfo.css
@@ -1,9 +1,9 @@
 .weightInfoNumber {
     margin-left: 1.25em;
+    overflow: hidden;
     border: 1.5px solid black;
     border-radius: 1rem;
-    overflow: hidden auto;
-    height: 100%;
+    height: 265px;
     text-align: right;
     text-shadow:
         -1px -1px 0 black,
