--- src/components/WeightInput.css
@@ -1,108 +1,0 @@
-.inputRow {
-    height: 100px;
-}
-
-.inputRow2 {
-    display: flex;
-    align-items: center;
-}
-
-#outlined-basic {
-    box-sizing: content-box;
-}
-
-.BarbellVisualization {
-    height: 100%;
-    display: flex;
-}
-
-.barbellImg {
-    height: 20%;
-}
-
-.plateImg {
-    max-width: 100%;
-    max-height: 100%;
-}
-
-.weightDisplayNumber {
-    position: absolute;
-    /* top: 50%; */
-    top: 100%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    
-    text-shadow:
-        -1px -1px 0 black,
-        1px -1px 0 black,
-        -1px 1px 0 black,
-        1px 1px 0 black;
-    color: white;
-    font-weight: 900;
-    display:inline-block;
-    font-size: 2em;
-    /* background-color: lightslategray;
-    padding: .5rem 1.5rem .5rem 1.5rem;
-    border-radius: 9999px; */
-}
-
-.flexCell {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 5px 0px 5px 0px;
-    height: 100%;
-}
-
-.bbIcon {
-    height: 100%;
-    padding-right: 1rem;
-}
-
-
-#weightInput {
-    box-sizing: border-box;
-    width: 100%;
-    height: calc(3em + 2px);
-    margin: 1em 0 1em;
-    padding: 1em;
-    border: 1.25px solid;
-    border-radius: .25em;
-    background: #fff;
-    resize: none;
-    outline: none;
-}
-#weightInput[required]:focus {
-    border: 2px solid #00bafa;
-}
-#weightInput[required]:focus + label[placeholder]:before {
-    color: #00bafa;
-}
-#weightInput[required]:focus + label[placeholder]:before, #weightInput[required]:valid + label[placeholder]:before {
-    transition-duration: .2s;
-    transform: translate(0,-1.5em) scale(.9,.9);
-}
-#weightInput[required]:invalid + label[placeholder][alt]:before {
-    content: attr(alt);
-}
-#weightInput[required] + label[placeholder] {
-    text-align: left;
-    display: block;
-    pointer-events: none;
-    line-height: 1.25em;
-    margin-top: calc(-3em - 2px);
-    margin-bottom: calc((3em - 1em) + 2px);
-}
-#weightInput[required] + label[placeholder]:before {
-    content: attr(placeholder);
-    display: inline-block;
-    margin: 0 calc(1em + 2px);
-    padding: 0 2px;
-    color: #898989;
-    white-space: nowrap;
-    transition: .3s ease-in-out;
-    background-image: linear-gradient(to bottom,#fff,#fff);
-    background-size: 100% 5px;
-    background-repeat: no-repeat;
-    background-position: center;
-}
