+++ src/components/WeightInput/WeightInput.scss
@@ -1,0 +1,57 @@
+.inputRow {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 1em 0 1em
+}
+
+#outlined-basic {
+    box-sizing: content-box;
+}
+
+.BarbellVisualization {
+    max-height: 250px;
+    display: flex;
+}
+
+.barbellImg {
+    height: 20%;
+}
+
+.plateImg {
+    max-width: 100%;
+    max-height: 100%;
+}
+
+.weightDisplayNumber {
+    position: absolute;
+    /* top: 50%; */
+    top: 100%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    
+    text-shadow:
+        -1px -1px 0 black,
+        1px -1px 0 black,
+        -1px 1px 0 black,
+        1px 1px 0 black;
+    color: white;
+    font-weight: 900;
+    display:inline-block;
+    font-size: 2em;
+    /* background-color: lightslategray;
+    padding: .5rem 1.5rem .5rem 1.5rem;
+    border-radius: 9999px; */
+}
+
+.flexCell {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 5px 0px 5px 0px;
+    height: 100%;
+}
+
+.bbIcon {
+    height: 25px;
+}
